Consolidate Windows CentralProcessor tick query into common base
queryProcessorCpuLoadTicks() was duplicated nearly verbatim between the
JNA and FFM subclasses, differing only in the perfmon driver call. Move
it, the memoized processorUtilityCounters supplier, and the initial-
counter bootstrap into the WindowsCentralProcessor base (which already
owns the related utility-counter state and processTickData), behind two
thin driver hooks: queryProcessorCapacityCounters() and
queryProcessorCounters(). The previously-abstract queryProcessorUtility-
Counters() becomes a concrete delegate. Eliminates ~70 duplicated lines;
the JNA and FFM subclasses keep only their native-specific code.
